moving ROS-specific KDL Lua code to rtt_kdl_conversions
